Create new files for ScenarioDesc
Move code to separate file.
Hide json writer details.
Fix unit tests to refer correct shader.
