+++ src/tests/resources/scenarios/test_spec_const/int_passthrough_glsl.json
@@ -8,7 +8,7 @@
       },
       {
         "dispatch_compute": {
-          "shader_ref": "shader",
+          "shader_ref": "int_shader",
           "rangeND": [
             1,
             1,
